ref: add authContext.Provider as React Component
