--- login-app/src/context/auth-context.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AuthContext = React.createContext({
-  isLoggedIn: false,
-});
-
-export default AuthContext;
