+++ login-app/src/index.jsx
@@ -4,10 +4,13 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import ThemeContext from "./context/ThemeContext";
+import { AuthContextProvider } from "./context/auth-context";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <ThemeContext>
-    <App />
+    <AuthContextProvider>
+      <App />
+    </AuthContextProvider>
   </ThemeContext>
 );
